feat: add new support module for NixOS channel change
There are many modules that depends on specific packages for certain configurations. So far, those dependencies are hard-coded to follow the stable channel (pkgs), but what if I want to switch NixOS to the unstable channel? At best, there will be a replication of resources wasting bulding time and storage; at worst there could even be packages collisions. This new module aims to introduce a new option that will default to `pkgs` (the stable channel) but can be easily set from a host's host-options.nix file to follow `pkgsUnstable` in case I want or need to source NixOS from nixos-unstable. A pretty cool feature, if you ask me.