refactor: the flake now loads the CPU architecture right from hardware-configuration.nix
Wow, just wow--we achieved a lot here! Thank you, Grok. :wink: :wink:

Like the title of the commit suggest, now the CPU architecture is dynamically inferred from the hardware-configureation.nix file.
This is HUGE because we don't need anymore to hardcode the system value for each host in flake.nix, ending up with the host CPU architecture being defined in two different places!

This refactor deserves a tag :)

Signed-off-by: Martin Cigorraga <cig0.github@tutanota.com>