chore: add git-crypt / nix-init
I checked out https://github.com/vlaci/git-agecrypt, as I'm transitioning from GPG to age to improve my workflows, but I don't feel is the right tool for me at this time. From the repository README:

/*
https://github.com/vlaci/git-agecrypt?tab=readme-ov-file#limitations

The following limitations can be easily improved upon, but they are not blockers for my use-case.

The application is started once for each file for every git operation. It can cause slowdown when the repository contains many encrypted files. A possible mitigation for this issue could be the implementation of the long-running process protocol but it is usable as it is for a couple of small files.

During encryption/decryption the whole file is loaded into memory. This can cause issues when encrypting large files.
*/

Signed-off-by: Martin Cigorraga <cig0.github@tutanota.com>