fix: hopefully fixed the way I was handling NH's FLAKE shell envvar
Previously, this shell envvar pointed to the NixOS store location of the
flake which, I belie, can cause issues when evaluating the flake, even
when explicityly pointing NH to the right directory or creating a
generation from there--the flake source code directory.

I believe this misconfiguration could be the cause of some obscure
isolated issues I've had in the past when building a new NixOS
generation.

I'll add a tag and monitor how the system behave.

Signed-off-by: Martin Cigorraga <cig0.github@tutanota.com>